# docs/swagger.yaml
basePath: /
definitions:
...
  apitoken.ApiTokenListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/apitoken.ApiToken'
        type: array
      total:
        type: integer
    type: object
...
  apitoken.CreateApiTokenRequest:
    properties:
      expires_at:
        type: string
      name:
        type: string
      routes:
        items:
          type: string
        type: array
    type: object
  apitoken.CreateApiTokenResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
...
  model.ApiError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  model.ListResponse-audit_AuditLogResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/audit.AuditLogResponse'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:9999
info:
  contact: {}
  description: Self-contained OAuth 2.0 / OpenID Connect (OIDC) Identity Provider
    built with Go and SQLite.
  title: Autentico OIDC
  version: "1.0"
paths:
...
  /admin/api/api-tokens:
    get:
      description: List all created API tokens
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apitoken.ApiTokenListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: List API Tokens
      tags:
      - apitokens
    post:
      consumes:
      - application/json
      description: Create a new API token with specific routes access
      parameters:
      - description: Create API Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apitoken.CreateApiTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apitoken.CreateApiTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Create API Token
      tags:
      - apitokens
  /admin/api/api-tokens/{id}:
    delete:
      description: Revoke an API token by ID
      parameters:
      - description: API Token ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Revoke API Token
      tags:
      - apitokens
...
  /admin/api/audit-logs:
    get:
      description: Returns a paginated list of audit events with optional filters,
        search, sorting, and date range.
      parameters:
      - description: Filter by event type
        in: query
        name: event
        type: string
      - description: Search actor_username, target_id, ip_address, detail
        in: query
        name: search
        type: string
      - default: created_at
        description: Sort field (created_at, event)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      - description: Date range start (ISO 8601)
        in: query
        name: created_at_from
        type: string
      - description: Date range end (ISO 8601)
        in: query
        name: created_at_to
        type: string
      - default: 100
        description: Max results per page (1–100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResponse-audit_AuditLogResponse'
      security:
      - AdminAuth: []
      summary: List audit log events
      tags:
      - admin-settings
...
  /admin/api/certificates:
    get:
      description: List all certificates
      parameters:
      - description: User ID
        in: query
        name: user
        type: string
      - description: Certificate Type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: List Certificates
      tags:
      - ca
    post:
      consumes:
      - application/json
      description: Generate a new certificate for a user
      parameters:
      - description: Generate Certificate Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ca.GenerateCertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ca.Certificate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Generate User Certificate
      tags:
      - ca
  /admin/api/certificates/{id}:
    delete:
      description: Revoke a certificate by ID
      parameters:
      - description: Certificate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Revoke Certificate
      tags:
      - ca
  /admin/api/certificates/{id}/bundle:
    get:
      description: Download a PKCS12 bundle for a user certificate
      parameters:
      - description: Certificate ID
        in: path
        name: id
        required: true
        type: string
      - description: Password for the bundle
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/x-pkcs12
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: Download User Certificate Bundle
      tags:
      - ca
  /admin/api/certificates/authorities:
    get:
      description: List all certificate authorities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      summary: List Certificate Authorities
      tags:
      - ca
...
  /admin/api/clients:
    get:
      description: Lists clients with server-side sorting, filtering, search, and
        pagination.
      parameters:
      - description: Sort field (created_at, client_name, client_id, client_type)
        in: query
        name: sort
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      - description: Search by client_name or client_id
        in: query
        name: search
        type: string
      - description: Filter by client type (confidential, public)
        in: query
        name: client_type
        type: string
      - description: Filter by active status (1, 0)
        in: query
        name: is_active
        type: string
      - default: 100
        description: Max results per page (1–100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResponse-client_ClientInfoResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.AuthErrorResponse'
      security:
      - AdminAuth: []
      summary: List clients with pagination
      tags:
      - admin-client
    post:
      consumes:
      - application/json
      description: Registers a new OAuth2/OIDC client. Also available at /oauth2/register
        (RFC 7591 Dynamic Client Registration).
      parameters:
      - description: Client registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.ClientCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/client.ClientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.AuthErrorResponse'
      security:
      - AdminAuth: []
      summary: Register a new OAuth2 client
      tags:
      - admin-client
  /admin/api/clients/{client_id}:
    delete:
      description: Deactivates (soft deletes) a registered client. Also available
        at /oauth2/register/{client_id}.
      parameters:
      - description: Client ID
        in: path
        name: client_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.AuthErrorResponse'
      security:
      - AdminAuth: []
      summary: Deactivate a client
      tags:
      - admin-client
    get:
      description: Retrieves information about a registered client. Also available
        at /oauth2/register/{client_id}.
      parameters:
      - description: Client ID
        in: path
        name: client_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.ClientInfoResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.AuthErrorResponse'
      security:
      - AdminAuth: []
      summary: Get client information
      tags:
      - admin-client
    put:
      consumes:
      - application/json
      description: Updates a registered client. Also available at /oauth2/register/{client_id}.
      parameters:
      - description: Client ID
        in: path
        name: client_id
        required: true
        type: string
      - description: Client update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/client.ClientUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.ClientInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.AuthErrorResponse'
      security:
      - AdminAuth: []
      summary: Update client information
      tags:
      - admin-client
  /admin/api/deletion-requests:
    get:
      parameters:
      - default: requested_at
        description: Sort by field
        enum:
        - requested_at
        - username
        - email
        in: query
        name: sort
        type: string
      - default: asc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: Search by username, email, or reason
        in: query
        name: search
        type: string
      - description: 'Filter: requested at or after (ISO 8601)'
        in: query
        name: requested_at_from
        type: string
      - description: 'Filter: requested at or before (ISO 8601)'
        in: query
        name: requested_at_to
        type: string
      - default: 100
        description: Max results per page (1–100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResponse-deletion_DeletionRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: List all pending deletion requests
      tags:
      - admin-deletion
  /admin/api/deletion-requests/{id}:
    delete:
      parameters:
      - description: Deletion request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Cancel (dismiss) a deletion request without deleting the user
      tags:
      - admin-deletion
  /admin/api/deletion-requests/{id}/approve:
    post:
      parameters:
      - description: Deletion request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Approve a deletion request — permanently deletes the user
      tags:
      - admin-deletion
  /admin/api/federation:
    get:
      description: Lists federation providers with server-side sorting, filtering,
        search, and pagination.
      parameters:
      - description: Sort field (name, issuer, client_id, sort_order, enabled, created_at)
        in: query
        name: sort
        type: string
      - default: asc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      - description: Search by name, issuer, or client_id
        in: query
        name: search
        type: string
      - description: Filter by enabled status (1, 0)
        in: query
        name: enabled
        type: string
      - default: 100
        description: Max results per page (1–100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResponse-federation_ProviderResponse'
      security:
      - AdminAuth: []
      summary: List federation providers
      tags:
      - admin-federation
    post:
      consumes:
      - application/json
      parameters:
      - description: Provider request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/federation.FederationProviderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminAuth: []
      summary: Create a federation provider
      tags:
      - admin-federation
  /admin/api/federation/{id}:
    delete:
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Delete a federation provider
      tags:
      - admin-federation
    get:
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/federation.ProviderResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Get a federation provider
      tags:
      - admin-federation
    put:
      consumes:
      - application/json
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      - description: Provider request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/federation.FederationProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Update a federation provider
      tags:
      - admin-federation
  /admin/api/groups:
    get:
      parameters:
      - description: Sort field (name, created_at, updated_at)
        in: query
        name: sort
        type: string
      - default: asc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      - description: Search across name and description
        in: query
        name: search
        type: string
      - default: 100
        description: Max results per page (1–100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResponse-group_GroupResponse'
      security:
      - AdminAuth: []
      summary: List all groups
      tags:
      - admin-groups
    post:
      consumes:
      - application/json
      parameters:
      - description: Group creation payload
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/group.GroupCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/group.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Create a new group
      tags:
      - admin-groups
  /admin/api/groups/{id}:
    delete:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Delete a group
      tags:
      - admin-groups
    get:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/group.GroupResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Get a group by ID
      tags:
      - admin-groups
    put:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Group update payload
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/group.GroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/group.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Update a group
      tags:
      - admin-groups
  /admin/api/groups/{id}/members:
    get:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/group.GroupMemberResponse'
            type: array
      security:
      - AdminAuth: []
      summary: List members of a group
      tags:
      - admin-groups
    post:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Member payload
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/group.GroupMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Add a user to a group
      tags:
      - admin-groups
  /admin/api/groups/{id}/members/{user_id}:
    delete:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - AdminAuth: []
      summary: Remove a user from a group
      tags:
      - admin-groups
  /admin/api/idp-sessions:
    get:
      description: Returns all active IdP sessions with server-side sorting, search,
        and pagination.
      parameters:
      - description: Sort field (last_activity_at, created_at)
        in: query
        name: sort
        type: string
      - default: asc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      - description: Search across username, email, and IP address
        in: query
        name: search
        type: string
      - default: 100
        description: Max results per page (1–100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        name: offset
        type: integer
      - description: Filter sessions created after (ISO 8601)
        in: query
        name: created_at_from
        type: string
      - description: Filter sessions created before (ISO 8601)
        in: query
        name: created_at_to
        type: string
      - description: Filter sessions active after (ISO 8601)
        in: query
        name: last_activity_at_from
        type: string
      - description: Filter sessions active before (ISO 8601)
        in: query
        name: last_activity_at_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResponse-idpsession_IdpSessionResponse'
      security:
      - AdminAuth: []
      summary: List IdP (device) sessions
      tags:
      - admin-sessions
...
  /admin/api/settings/export:
    get:
      description: Export all settings as a JSON file (smtp_password excluded).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/appsettings.settingsExport'
      security:
      - AdminAuth: []
      summary: Export settings
      tags:
      - admin-settings
...
securityDefinitions:
  AdminAuth:
    description: Admin access token — requires admin role and autentico-admin audience.
      Type "Bearer" followed by your token.
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    description: User access token — any valid bearer token. Type "Bearer" followed
      by your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...